- Fixed code contracts issues

+++ Source code/Theme/Converters/PositiveToNegativeConverter.cs
@@ -16,9 +16,9 @@ namespace Elysium.Theme.Converters
             if (value is Int16)
                 return (Int16)((Int16)value * -1);
             if (value is Int32)
-                return (Int32)((Int32)value * -1);
+                return (Int32)value * -1;
             if (value is Int64)
-                return (Int64)((Int64)value * -1);
+                return (Int64)value * -1;
             if (value is Single)
                 return (Single)value * -1f;
             if (value is Double)
